[MISC] move type_traits/range to core/range
