+++ include/seqan3/alignment/matrix/detail/combined_score_and_trace_matrix.hpp
@@ -17,7 +17,7 @@
 #include <seqan3/alignment/matrix/detail/affine_cell_proxy.hpp>
 #include <seqan3/alignment/matrix/detail/matrix_coordinate.hpp>
 #include <seqan3/core/concept/core_language.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/zip.hpp>
 
 namespace seqan3::detail
