+++ include/seqan3/alignment/matrix/detail/two_dimensional_matrix.hpp
@@ -21,7 +21,7 @@
 #include <seqan3/alignment/matrix/detail/two_dimensional_matrix_iterator_base.hpp>
 #include <seqan3/core/detail/deferred_crtp_base.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 
 namespace seqan3::detail
 {
