+++ include/seqan3/contrib/stream/bgzf_stream_util.hpp
@@ -24,12 +24,13 @@
 #error "This file cannot be used when building without GZip-support."
 #endif  // SEQAN3_HAS_ZLIB
 
-#include <seqan3/core/bit_manipulation.hpp>
-#include <seqan3/core/type_traits/range.hpp>
-#include <seqan3/io/detail/magic_header.hpp>
-#include <seqan3/io/exception.hpp>
 #include <seqan3/std/algorithm>
 #include <seqan3/std/span>
+
+#include <seqan3/core/bit_manipulation.hpp>
+#include <seqan3/core/range/type_traits.hpp>
+#include <seqan3/io/detail/magic_header.hpp>
+#include <seqan3/io/exception.hpp>
 
 namespace seqan3::contrib
 {
