+++ include/seqan3/core/detail/debug_stream_range.hpp
@@ -18,7 +18,7 @@
 #include <seqan3/alphabet/adaptation/uint.hpp>
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/core/detail/debug_stream_type.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/concept.hpp>
 
 namespace seqan3::detail
