+++ include/seqan3/io/alignment_file/format_bam.hpp
@@ -26,7 +26,7 @@
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/detail/debug_stream_optional.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/io/alignment_file/detail.hpp>
 #include <seqan3/io/alignment_file/format_sam_base.hpp>
 #include <seqan3/io/alignment_file/header.hpp>
