+++ include/seqan3/io/alignment_file/format_sam.hpp
@@ -23,7 +23,7 @@
 #include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/io/alignment_file/detail.hpp>
 #include <seqan3/io/alignment_file/format_sam_base.hpp>
 #include <seqan3/io/alignment_file/header.hpp>
