+++ include/seqan3/range/views/take_until.hpp
@@ -19,8 +19,8 @@
 #include <seqan3/std/type_traits>
 
 #include <seqan3/core/detail/iterator_traits.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/core/semiregular_box.hpp>
-#include <seqan3/core/type_traits/range.hpp>
 #include <seqan3/io/exception.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/views/detail.hpp>
