+++ include/seqan3/search/fm_index/concept.hpp
@@ -16,7 +16,7 @@
 
 #include <sdsl/suffix_arrays.hpp>
 
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/container/concept.hpp>
