+++ test/include/seqan3/test/performance/naive_kmer_hash.hpp
@@ -12,12 +12,13 @@
 
 #pragma once
 
+#include <seqan3/std/ranges>
+
 #include <range/v3/view/sliding.hpp>
 
 #include <seqan3/alphabet/concept.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/detail.hpp>
-#include <seqan3/std/ranges>
 
 namespace seqan3::detail
 {
