+++ test/snippet/utility/type_pack/pack_traits_transform.cpp
@@ -1,8 +1,8 @@
 #include <list>
 #include <vector>
 
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/utility/type_pack/traits.hpp>
-#include <seqan3/core/type_traits/range.hpp>
 
 int main()
 {
