+++ test/unit/range/views/repeat_test.cpp
@@ -7,14 +7,15 @@
 
 #include <gtest/gtest.h>
 
+#include <seqan3/std/algorithm>
+#include <seqan3/std/ranges>
+
 #include <seqan3/core/detail/iterator_traits.hpp>
-#include <seqan3/core/type_traits/range.hpp>
+#include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/persist.hpp>
 #include <seqan3/range/views/repeat.hpp>
 #include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
-#include <seqan3/std/algorithm>
-#include <seqan3/std/ranges>
 #include <seqan3/test/expect_range_eq.hpp>
 
 TEST(repeat_view, deduction_guide)
